Update readvertiser functionality to reconcile shoot cluster endpoints
currently the readvertiser updates the deployment of the api-server with
the new advertise address seen from the AWS ELB, this causes the API
server to restart frequently. This commit, updates this functionality by
allowing the re-advertiser to edit the cluster endpoint directly
assuming the api-server --endpoint-reconciler-type is set to none. This
way, less api-restarts will be acheived.